Separate filters to their own files
