+++ API/ChallengeList.js
@@ -2,10 +2,16 @@ import APIAdapter from "./APIadapter.js";
 
 export default class ChallengeList {
     challenges;
+    filters;
+
     async render() {    
         const api = new APIAdapter();
-        const challenges = await api.loadChallenges();
-        
+        this.challenges = await api.loadChallenges();
+
+        this.filters = new Filters(this);
+
+
+
         const ul = document.createElement('ul');
         
         // for (let i = 0; i < this.challenges.length; i++) {
