+++ FILTERS/RatingFilter.js
@@ -1,0 +1,31 @@
+export class RatingFilter{
+    constructor(list){
+        this.list = list;
+    }
+
+    challengeMatch(challenge){
+
+        // Kolla ifall challenge matchar emot rating given av user.
+        // document.querySelector... är placeholder i nuläget
+        // man skulle kunna skriva kod som lägger till id/class på just den stjärnan som clickas på samt
+        // tar bort på de andra  
+
+        if(challenge.data.rating >= document.querySelector('#ratingMin').value && challenge.data.rating <= document.querySelector('#ratingMax').value){
+            return true;
+        } else {
+            return false;
+        }
+    }
+
+    render(){
+        //https://medium.com/codex/create-a-star-rating-widget-with-css-in-9-steps-fe323352dba4
+
+        // Fördel med att använda radioknappar är att de har attributet value som vi kan sätta
+        // till en siffra och på så sätt kontrollera mot .rating i challenge
+
+        // Här ska en rating input renderas i DOM. förslag att använda radioknappar som blir 
+        // dependent av varandra. dvs är nr 3 av 5 selected så tilldelas 1-3 samma styling. 
+        // Skit i halva stjärnor.
+    }
+
+}
